changes based on code review. removed commented code, log claims for newly added user, moved extension method to better place
